Compute our JavaComponent data on the fly.
This avoids pesky race conditions when computing our build classpath triggers
the resolution of other projects which then don't immediately have their own
metadata ready.

There can still be glitches the very first time, but this is tolerable for now
and I'm not sure I want to rewrite everything to be reactive or use futures.